re-org voicemail and sms folder structure and fix imports
