+++ vite.config.ts
@@ -2,10 +2,16 @@ import { defineConfig } from 'vite';
 import tailwindcss from '@tailwindcss/vite';
 
 import react from '@vitejs/plugin-react-swc';
+import path from 'path'
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react(), tailwindcss()],
+  resolve: {
+    alias: {
+      '@client': path.resolve(__dirname, './src/client')
+    }
+  },
   server: {
     proxy: {
       '/api': {
